Added most of the needed functionality, very roughly
The login and sign up feature is implemented.
Events are pulled from the Firebase Database.
Events are displayed on the map.
The list is populated with event cards.
The detail view shows all the event information.
The detail view allows the event creator to edit their events.
Users can add events.